refactor location of functions and questions
